% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_march.R
\name{csw}
\alias{csw}
\title{Renge estimation by constant specific work hypothesis}
\usage{
csw(data, control = list())
}
\arguments{
\item{data}{A dataframe of observations}

\item{control}{A list of flight control parameters}
}
\description{
Renge estimation by constant specific work hypothesis
}
\author{
Brian Masinde
}
