% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curlew_sandpiper.R
\docType{data}
\name{curlew_sandpiper}
\alias{curlew_sandpiper}
\title{Curlew Sandpiper Data}
\format{
A dataframe with 60 rows and 9 columns:

\describe{
   \item{day}{Day of observation of bird}
   \item{name}{Unique name of individual bird}
   \item{empty mass}{Body mass in Kg. of the birds}
   \item{fat.mass}{Mass of fat that is consumable as fuel in Kg}
   \item{Wing.area}{Area of both wing projected on a flat surface in meters
   squared}
   \item{Wing.span}{Length of wings spread out in meters}
   \item{order}{Order of the species (passerine = 1 vs non-passerine = 2)}
   \item{Muscle.mass}{Mass in Kg. of flight muscles calculated as 0.17 fraction 
   of body mass}
   

}
}
\usage{
curlew_sandpiper
}
\description{
A sample of 12 individuals observed in lab conditions over five days at the 
University of Gdansk. Birds were fed ad-libitum with mealworm larvae Tenebrio molitor. 
To determine the maximum body mass gain sandpipers were weighed every day in the morning
}
\keyword{datasets}
