% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_1.R
\name{.breguet}
\alias{.breguet}
\title{Method 1 on practical range calculation based on Breguets equations}
\usage{
.breguet(bodyMass, wingSpan, fatMass, ordo, wingArea, ctrl)
}
\arguments{
\item{bodyMass}{all up mass}

\item{wingSpan}{wing span of bird in metres}

\item{fatMass}{fat mass of bird}

\item{ordo}{Passerine (1) or non-passerine (2)}

\item{wingArea}{area of wing}

\item{ctrl}{A list of re-definition of constants (i.e *airDensity*,
*consume*, *enegry e*, *mechanical efficiency n*).}
}
\value{
List with range (in km), constants used and fat fraction
}
\description{
Method 1 on practical range calculation based on Breguets equations
}
\author{
Brian Masinde
}
