% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_2.R
\name{.breguet_adj}
\alias{.breguet_adj}
\title{Method 2 practical range calculation based on Breguets equations with mean
 of effective lift: drag ratio}
\usage{
.breguet_adj(bodyMass, wingSpan, fatMass, ordo, wingArea, ctrl)
}
\arguments{
\item{bodyMass}{All up mass. Including fuel, crop contents and equipment in Kg}

\item{wingSpan}{Wing span in metres}

\item{fatMass}{Fat mass in kg (fuel)}

\item{ordo}{Passerine (1) or non-passerine (2)}

\item{wingArea}{Wing area}

\item{ctrl}{A list of re-definition of constants (i.e *airDensity*,
*consume*, *enegry e*, *mechanical efficiency n*).}
}
\description{
Practical range estimation using Breguet equation for fixed wing
             with crude adjustments. Mean lift:drag ratio between start and
             end of flight is used as proxy for engine burn.
}
\author{
Brian Masinde
}
