% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_match.R
\name{.colnames.match}
\alias{.colnames.match}
\title{Identify and rename columns in Data}
\usage{
.colnames.match(data)
}
\value{
data returns processed data
}
\description{
Identify and rename columns in Data
}
\author{
Brian Masinde
}
