% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_table2.R
\name{.gen.table2}
\alias{.gen.table2}
\title{load table 2 generate. Description in Pennycuick earlier version}
\usage{
.gen.table2()
}
\value{
table2
}
\description{
Pennycuick's table II aids in calculation of D factor for finding
             lift:drag ratio. C factor for finding power required at maximum
             range speed and B a ratio of maximum range speed and minimum
             power speed.
}
\author{
Brian Masinde
}
