% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_to_data.R
\name{.pathToData}
\alias{.pathToData}
\title{path to data}
\usage{
.pathToData(fileArgs = list(), ...)
}
\value{
data processed data from path
}
\description{
Handles arguments for path to data.
}
\author{
Brian Masinde
}
