% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesser_whitethroats.R
\docType{data}
\name{lesser_whitethroats}
\alias{lesser_whitethroats}
\title{Lesser Whitethroats Data}
\format{
A dataframe with 84 rows and 9 columns:

\describe{
   \item{Region}{Region of ringing station where bird was captured for measurements}
   \item{body mass}{Body mass in Kg. of the birds}
   \item{Fat.mass}{Mass of fat that is consumable as fuel in Kg}
   \item{Wing.span}{Length of wings spread out in meters}
   \item{Wing.area}{Area of both wing projected on a flat surface in meters
   squared}
   \item{Muscle.mass}{Mass in Kg. of flight muscles calculated as 0.17 fraction 
   of body mass}
   \item{taxon}{Order of the species (passerine = 1 vs non-passerine = 2)}

}
}
\usage{
lesser_whitethroats
}
\description{
A subset of data on Lesser Whitethroats captured during autumn migration
across several ringing stations in along the southern eastern migration
flyway  (See Ożarowska, A. 2015).
}
\keyword{datasets}
