% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopover_mass_calculator.R
\name{stopover.mass.calculator}
\alias{stopover.mass.calculator}
\title{Stopover mass calculator}
\usage{
stopover.mass.calculator(
  file,
  header = TRUE,
  sep = ",",
  quote = "\\"",
  dec = ".",
  fill = TRUE,
  comment.char = "",
  ...,
  data = NULL,
  duration = 1L
)
}
\arguments{
\item{file}{The name of the file which the data are to read from}

\item{header}{Logical. If TRUE use first row as column headers}

\item{sep}{separator}

\item{quote}{The set of quoting characters. see read.csv}

\item{dec}{The character used in the file for decimal points}

\item{fill}{See read.csv}

\item{comment.char}{For more details see read.csv}

\item{...}{further arguments see read.csv}

\item{data}{A data frame}

\item{duration}{Days spent at stopover}
}
\value{
body mass, fat mass, fat fraction,
}
\description{
During stop-overs birds replenish fat mass. Using simplifications
from Lindstöm 1991.
}
