% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight_simulation.R
\name{flysim}
\alias{flysim}
\title{Implementation of the methods 1 and 2 base on Pennycuick. Both use Breguet's
equations for lift drag ratio calculation.}
\usage{
flysim(data, method, ctrl)
}
\arguments{
\item{data}{A data frame or a list (for a single bird observation). See example}

\item{method}{Two methods are currently available: "breguet" and "breguet_adj"}

\item{ctrl}{A list of re-definition of constants (*airDensity*,
*consume*, *enegry e*, *mechanical efficiency n*).}
}
\value{
S3 class object with range estimates based on methods defined and
       constants
}
\description{
Implementation of the methods 1 and 2 base on Pennycuick. Both use Breguet's
equations for lift drag ratio calculation.
}
\examples{
flysim(data = birds, ctrl = list(energy = 3.89*10^7))
flysim(data = birds, method = "breguet", ctrl = list(airDensity = 0.905))
flysim(data = birds, method = "breguet_adj", ctrl = list(airDensity = 0.905))
}
\author{
Brian Masinde
}
